+++ core/src/main/kotlin/net/corda/core/internal/GlobalProperties.kt
@@ -2,11 +2,11 @@ package net.corda.core.internal
 
 import net.corda.core.node.NetworkParameters
 
-object NodeProperties {
+object GlobalProperties {
     private var _networkParameters: NetworkParameters? = null
 
     var networkParameters: NetworkParameters
-        get() = _networkParameters ?: throw IllegalArgumentException("Property 'networkParameters' has not been initialised.")
+        get() = checkNotNull(_networkParameters) { "Property 'networkParameters' has not been initialised." }
         set(value) {
             _networkParameters = value
         }