+++ node/src/main/kotlin/net/corda/node/internal/AbstractNode.kt
@@ -14,7 +14,7 @@ import net.corda.core.identity.CordaX500Name
 import net.corda.core.identity.Party
 import net.corda.core.identity.PartyAndCertificate
 import net.corda.core.internal.FlowStateMachine
-import net.corda.core.internal.NodeProperties.networkParameters
+import net.corda.core.internal.GlobalProperties.networkParameters
 import net.corda.core.internal.VisibleForTesting
 import net.corda.core.internal.concurrent.map
 import net.corda.core.internal.concurrent.openFuture