+++ docs/source/changelog.rst
@@ -6,9 +6,6 @@ from the previous milestone release.
 
 UNRELEASED
 ----------
-* Introduced max transaction size limit on transactions. The max transaction size parameter is set by the compatibility zone
-  operator. The parameter is distributed to Corda nodes by network map service as part of the ``NetworkParameters``.
-
 * Removed blacklisted word checks in Corda X.500 name to allow "Server" or "Node" to be use as part of the legal name.
 
 * Separated our pre-existing Artemis broker into an RPC broker and a P2P broker.
@@ -77,6 +74,9 @@ UNRELEASED
 
    * Moved ``NodeInfoSchema`` to internal package as the node info's database schema is not part of the public API. This
      was needed to allow changes to the schema.
+
+   * Introduced max transaction size limit on transactions. The max transaction size parameter is set by the compatibility zone
+     operator. The parameter is distributed to Corda nodes by network map service as part of the ``NetworkParameters``.
 
 * Support for external user credentials data source and password encryption [CORDA-827].